@@ -7,7 +7,8 @@ import { hasProperty, isEmpty } from '../utils/utils';
 import { MISSING_PRIVATE_BETA_SERVICE_ERROR } from '../utils/error';
 import { getServicesForSelect } from '../utils/serviceUtils';
 import { UserType } from '../utils/UserType';
-import { PRIVATE_BETA_ROLE } from '../utils/serviceUtils';
+import { Service } from '../interfaces/Service';
+import { Role } from '../interfaces/Role';
 
 @autobind
 export class AddPrivateBetaServiceController extends RootController {
@@ -15,7 +16,8 @@ export class AddPrivateBetaServiceController extends RootController {
   public async post(req: AuthedRequest, res: Response) {
     const allServices = await req.scope.cradle.api.getAllServices();
     const fields = req.body;
-    const privateBetaServices = getServicesForSelect(allServices);
+    const rolesMap = await this.getRolesMap(req);
+    const privateBetaServices = getServicesForSelect(allServices, rolesMap);
     const user = {
       email: fields._email,
       forename: fields._forename,
@@ -28,15 +30,13 @@ export class AddPrivateBetaServiceController extends RootController {
         error: { privateBeta : { message: MISSING_PRIVATE_BETA_SERVICE_ERROR } }
       });
     }
-
-    const selectedService = allServices.find(service => service.label === fields.service);
-    const privateBetaRole = selectedService.onboardingRoles.find(element => element.includes(PRIVATE_BETA_ROLE));
+    const rolesToAdd = await this.getRolesToRegisterUser(req, allServices, fields.service);
 
     return req.scope.cradle.api.registerUser({
       email: fields._email,
       firstName: fields._forename,
       lastName: fields._surname,
-      roles: [UserType.Citizen, privateBetaRole]
+      roles: rolesToAdd
     })
       .then (() => {
         return super.post(req, res, 'add-user-completion');
@@ -47,5 +47,25 @@ export class AddPrivateBetaServiceController extends RootController {
           error: { userRegistration : { message: error } }
         });
       });
+  }
+
+  private async getRolesToRegisterUser(req: AuthedRequest, allServices: Service[], serviceField: string): Promise<string[]> {
+    const selectedService = allServices.find(service => service.label === serviceField);
+    const rolesToAdd: string[] = [UserType.Citizen];
+    const rolesMap = await this.getRolesMap(req);
+
+    selectedService.onboardingRoles
+      .filter(r => rolesMap.has(r))
+      .forEach(r => rolesToAdd.push(rolesMap.get(r).name));
+    return rolesToAdd;
+  }
+
+  private async getRolesMap(req: AuthedRequest): Promise<Map<string, Role>> {
+    const allRoles = await req.scope.cradle.api.getAllRoles();
+    const rolesMap = new Map(allRoles
+      .filter(role => role !== undefined)
+      .map(role => [role.id, role])
+    );
+    return rolesMap;
   }
 }
