@@ -1,5 +1,5 @@
 export interface SelectItem {
   value: string | number;
   text: string;
-  selected: boolean;
+  selected?: boolean;
 }
