@@ -18,6 +18,9 @@ import { IdamAuth } from '../../app/idam-auth/IdamAuth';
 import config from 'config';
 import { UserEditController } from '../../controllers/UserEditController';
 import { AccessibilityStatementController } from '../../controllers/AccessibilityStatementController';
+import { GenerateReportController } from '../../controllers/GenerateReportController';
+import { ReportsHandler } from '../../app/reports/ReportsHandler';
+import { DownloadReportController } from '../../controllers/DownloadReportController';
 import { AddPrivateBetaServiceController } from '../../controllers/AddPrivateBetaServiceController';
 
 /**
@@ -31,6 +34,7 @@ export class Container {
       telemetryClient: asValue(defaultClient),
       exposeErrors: asValue(app.locals.env === 'development'),
       featureFlags: asValue(new FeatureFlags(new LaunchDarkly())),
+      reportGenerator: asValue(new ReportsHandler(logger, defaultClient)),
       userOptionController: asClass(UserOptionController),
       addUserController: asClass(AddUserController),
       addUserDetailsController: asClass(AddUserDetailsController),
@@ -42,7 +46,9 @@ export class Container {
       userActionsController: asClass(UserActionsController),
       userDeleteController: asClass(UserDeleteController),
       userSuspendController: asClass(UserSuspendController),
-      accessibilityStatementController: asClass(AccessibilityStatementController)
+      accessibilityStatementController: asClass(AccessibilityStatementController),
+      generateReportController: asClass(GenerateReportController),
+      downloadReportController: asClass(DownloadReportController)
     });
 
     /**
