@@ -8,8 +8,7 @@ import '../../main/utils/utils';
 import {config as testConfig} from '../config';
 import * as Assert from 'assert';
 import {randomData} from './shared/random-data';
-import {BETA_ADD} from '../../main/app/feature-flags/flags';
-import {PRIVATE_BETA_ROLE} from '../../main/utils/serviceUtils';
+import {BETA_ADD, GAMMA_PRIVATE_BETA} from '../../main/app/feature-flags/flags';
 import {UserType} from '../../main/utils/UserType';
 
 Feature('Register New User');
@@ -19,7 +18,7 @@ const ASSIGNABLE_CHILD_ROLE1 = randomData.getRandomRole();
 const ASSIGNABLE_CHILD_ROLE2 = randomData.getRandomRole();
 const DASHBOARD_USER_EMAIL = randomData.getRandomEmailAddress();
 const SERVICE_WITH_PRIVATE_BETA = randomData.getRandomRole();
-const PRIVATE_BETA_USER_ROLE = SERVICE_WITH_PRIVATE_BETA + '-' + PRIVATE_BETA_ROLE;
+const PRIVATE_BETA_USER_ROLE = SERVICE_WITH_PRIVATE_BETA + '-beta-role';
 
 const OAUTH_REDIRECT_URI = 'http://test.com/oauth2/callback';
 
@@ -62,7 +61,7 @@ Scenario('I as a user should be able to register new support user',
     I.waitForText('User registered');
 
     const response = await I.extractUrlFromNotifyEmail(registerUserEmail);
-    const activationParams = response.match(/token=(.*?)&code=([a-zA-Z0-9\\-]+)/);
+    const activationParams = response.body.match(/token=(.*?)&code=([a-zA-Z0-9\\-]+)/);
     const token = activationParams[1];
     const code = activationParams[2];
     await I.activateUserAccount(code, token);
@@ -80,7 +79,7 @@ Scenario('I as a user should be able to register new support user',
   }).tag('@CrossBrowser');
 
 Scenario('I as a user should be able to register new private beta citizen user',
-  {featureFlags: [BETA_ADD]},
+  {featureFlags: [BETA_ADD, GAMMA_PRIVATE_BETA]},
   async ({I}) => {
     const registerUserEmail = randomData.getRandomEmailAddress();
     I.loginAs(DASHBOARD_USER_EMAIL, testConfig.PASSWORD);
@@ -106,7 +105,7 @@ Scenario('I as a user should be able to register new private beta citizen user',
     I.waitForText('User registered');
 
     const response = await I.extractUrlFromNotifyEmail(registerUserEmail);
-    const activationParams = response.match(/token=(.*?)&code=([a-zA-Z0-9\\-]+)/);
+    const activationParams = response.body.match(/token=(.*?)&code=([a-zA-Z0-9\\-]+)/);
     const token = activationParams[1];
     const code = activationParams[2];
     await I.activateUserAccount(code, token);
